# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Compurge Timestamps API
  version: 0.1.0
  description: >
    Backend API for generating timestamped clip ranges from football event data.
    The backend parses event data, applies filters, resolves timeline offsets,
    and returns clip ranges. It does not render video.

servers:
  - url: http://localhost:8080

paths:
  /timestamps:
    post:
      summary: Generate clip ranges from uploaded event data
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - eventData
              properties:
                eventData:
                  type: string
                  format: binary
                requestId:
                  type: string
                period:
                  type: string
                  example: FirstHalf
                eventType:
                  type: string
                  example: Pass
                outcomeType:
                  type: string
                  example: Successful
                minMinute:
                  type: integer
                maxMinute:
                  type: integer
                minX:
                  type: number
                  format: double
                maxX:
                  type: number
                  format: double
                minY:
                  type: number
                  format: double
                maxY:
                  type: number
                  format: double
                minEndX:
                  type: number
                  format: double
                maxEndX:
                  type: number
                  format: double
                minEndY:
                  type: number
                  format: double
                maxEndY:
                  type: number
                  format: double
                minXT:
                  type: number
                  format: double
                maxXT:
                  type: number
                  format: double
                minProgPass:
                  type: number
                  format: double
                maxProgPass:
                  type: number
                  format: double
                minProgCarry:
                  type: number
                  format: double
                maxProgCarry:
                  type: number
                  format: double
                preRollSeconds:
                  type: number
                  format: double
                  example: 2
                postRollSeconds:
                  type: number
                  format: double
                  example: 3
                timelineOffset:
                  type: number
                  format: double
                  example: 5
                firstHalfOffset:
                  type: number
                  format: double
                  example: 5
                secondHalfOffset:
                  type: number
                  format: double
                  example: 120
                mergeGapSeconds:
                  type: number
                  format: double
                  example: 0
      responses:
        "200":
          description: Clip ranges generated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimestampResponse"
        "400":
          description: Invalid request
        "429":
          description: Server busy
        "500":
          description: Internal server error

  /parse-preview:
    post:
      summary: Parse event data and return a limited preview event list
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - eventData
              properties:
                eventData:
                  type: string
                  format: binary
                requestId:
                  type: string
                limit:
                  type: integer
                  example: 5
      responses:
        "200":
          description: Preview events returned successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PreviewResponse"
        "400":
          description: Invalid request
        "429":
          description: Server busy
        "500":
          description: Internal server error

  /timestamps/json:
    post:
      summary: Generate clip ranges from raw JSON input
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: "#/components/schemas/TimestampJSONCSVRequest"
                - $ref: "#/components/schemas/TimestampJSONEventsRequest"
            examples:
              csvText:
                summary: Raw CSV text payload
                value:
                  requestId: player_a_first_half_passes
                  eventDataCsv: |-
                    minute,second,period,type,outcomeType
                    0,24,FirstHalf,Pass,Successful
                    1,45,FirstHalf,Pass,Successful
                  period: FirstHalf
                  eventType: Pass
                  outcomeType: Successful
                  minXT: 0.01
                  minProgPass: 10
                  minEndX: 80
                  minEndY: 18
                  maxEndY: 62
                  preRollSeconds: 2
                  postRollSeconds: 3
                  timelineOffset: 5
                  firstHalfOffset: 5
                  secondHalfOffset: 120
                  mergeGapSeconds: 0
              structuredEvents:
                summary: Structured events payload
                value:
                  events:
                    - minute: 0
                      second: 24
                      period: FirstHalf
                      eventType: Pass
                      outcomeType: Successful
                    - minute: 1
                      second: 45
                      period: FirstHalf
                      eventType: Pass
                      outcomeType: Successful
                  period: FirstHalf
                  eventType: Pass
                  minXT: 0.01
                  minProgPass: 10
                  minEndX: 80
                  preRollSeconds: 2
                  postRollSeconds: 3
                  firstHalfOffset: 5
                  secondHalfOffset: 120
      responses:
        "200":
          description: Clip ranges generated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimestampResponse"
        "400":
          description: Invalid request
        "500":
          description: Internal server error

  /parse-preview/json:
    post:
      summary: Parse JSON event input and return a limited preview event list
      requestBody:
        required: true
        content:
          application/json:
            schema:
              oneOf:
                - $ref: "#/components/schemas/PreviewJSONCSVRequest"
                - $ref: "#/components/schemas/PreviewJSONEventsRequest"
            examples:
              csvTextPreview:
                summary: Raw CSV text preview payload
                value:
                  limit: 3
                  eventDataCsv: |-
                    minute,second,period,type,outcomeType
                    0,24,FirstHalf,Pass,Successful
                    1,45,FirstHalf,Carry,Successful
                    46,10,SecondHalf,Pass,Successful
              structuredEventsPreview:
                summary: Structured events preview payload
                value:
                  limit: 2
                  events:
                    - minute: 0
                      second: 24
                      period: FirstHalf
                      eventType: Pass
                      outcomeType: Successful
                    - minute: 1
                      second: 45
                      period: FirstHalf
                      eventType: Carry
                      outcomeType: Successful
      responses:
        "200":
          description: Preview events returned successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PreviewResponse"
        "400":
          description: Invalid request
        "500":
          description: Internal server error

components:
  schemas:
    ClipRange:
      type: object
      properties:
        eventTime:
          type: number
          format: double
        resolvedEventTime:
          type: number
          format: double
        resolvedStartTime:
          type: number
          format: double
        resolvedEndTime:
          type: number
          format: double
        sourceEventType:
          type: string
        sourceOutcomeType:
          type: string

    TimestampResponse:
      type: object
      properties:
        requestId:
          type: string
        clipRanges:
          type: array
          items:
            $ref: "#/components/schemas/ClipRange"

    PreviewEvent:
      type: object
      properties:
        minute:
          type: integer
        second:
          type: number
          format: double
        period:
          type: string
        matchSecond:
          type: number
          format: double
        playerName:
          type: string
        teamName:
          type: string
        eventType:
          type: string
        outcomeType:
          type: string
        xT:
          type: number
          format: double

    PreviewResponse:
      type: object
      properties:
        requestId:
          type: string
        events:
          type: array
          items:
            $ref: "#/components/schemas/PreviewEvent"

    EventPayload:
      type: object
      required:
        - minute
        - second
        - period
      properties:
        minute:
          type: integer
        second:
          type: number
          format: double
        period:
          type: string
          example: FirstHalf
        playerId:
          type: string
        playerName:
          type: string
        teamId:
          type: string
        teamName:
          type: string
        eventType:
          type: string
        outcomeType:
          type: string
        progPass:
          type: number
          format: double
        progCarry:
          type: number
          format: double
        xT:
          type: number
          format: double
        x:
          type: number
          format: double
        y:
          type: number
          format: double
        endX:
          type: number
          format: double
        endY:
          type: number
          format: double

    TimestampJSONBase:
      type: object
      properties:
        limit:
          type: integer
          example: 50
        requestId:
          type: string
          example: player_a_first_half_passes
        period:
          type: string
          example: FirstHalf
        eventType:
          type: string
          example: Pass
        outcomeType:
          type: string
          example: Successful
        minMinute:
          type: integer
          example: 0
        maxMinute:
          type: integer
          example: 10
        minX:
          type: number
          format: double
          example: 25
        maxX:
          type: number
          format: double
          example: 40
        minY:
          type: number
          format: double
          example: 18
        maxY:
          type: number
          format: double
          example: 62
        minEndX:
          type: number
          format: double
          example: 80
        maxEndX:
          type: number
          format: double
          example: 100
        minEndY:
          type: number
          format: double
          example: 18
        maxEndY:
          type: number
          format: double
          example: 62
        minXT:
          type: number
          format: double
          example: 0.01
        maxXT:
          type: number
          format: double
          example: 0.2
        minProgPass:
          type: number
          format: double
          example: 10
        maxProgPass:
          type: number
          format: double
          example: 50
        minProgCarry:
          type: number
          format: double
          example: 5
        maxProgCarry:
          type: number
          format: double
          example: 30
        preRollSeconds:
          type: number
          format: double
          example: 2
        postRollSeconds:
          type: number
          format: double
          example: 3
        timelineOffset:
          type: number
          format: double
          example: 5
        firstHalfOffset:
          type: number
          format: double
          example: 5
        secondHalfOffset:
          type: number
          format: double
          example: 120
        mergeGapSeconds:
          type: number
          format: double
          example: 0

    TimestampJSONCSVRequest:
      allOf:
        - $ref: "#/components/schemas/TimestampJSONBase"
        - type: object
          required:
            - eventDataCsv
          properties:
            eventDataCsv:
              type: string
              example: |-
                minute,second,period,type,outcomeType
                0,24,FirstHalf,Pass,Successful
                1,45,FirstHalf,Pass,Successful

    TimestampJSONEventsRequest:
      allOf:
        - $ref: "#/components/schemas/TimestampJSONBase"
        - type: object
          required:
            - events
          properties:
            events:
              type: array
              items:
                $ref: "#/components/schemas/EventPayload"

    PreviewJSONBase:
      type: object
      properties:
        requestId:
          type: string
        limit:
          type: integer
          example: 50

    PreviewJSONCSVRequest:
      allOf:
        - $ref: "#/components/schemas/PreviewJSONBase"
        - type: object
          required:
            - eventDataCsv
          properties:
            eventDataCsv:
              type: string

    PreviewJSONEventsRequest:
      allOf:
        - $ref: "#/components/schemas/PreviewJSONBase"
        - type: object
          required:
            - events
          properties:
            events:
              type: array
              items:
                $ref: "#/components/schemas/EventPayload"